switch to retry-based interruptible training, single file
